+++ CHANGES.md
@@ -6,7 +6,8 @@ releases are available on [Anaconda.org](https://anaconda.org/conda-forge/gettsi
 
 ## Unpublished
 
-- {gh}`577` Implement pension for reduced earning capacity ({ghuser}`nafetsk`)
+- {gh}`577` Implement Erwerbsminderungsrente (public disability insurance)
+  ({ghuser}`nafetsk`)
 - {gh}`634` Implement pension for unemployed ({ghuser}`TeBackh`)
 - {gh}`632` Zugangsfaktor calculation with monthly precision ({ghuser}`TeBackh`).
 - {gh}`624` Don't create functions for other time units if this leads to a cycle in the